factory/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
jsonschema==4.21.1
python-jose[cryptography]==3.3.0
requests==2.31.0
python-multipart==0.0.9
